requirements.txt
flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.1
supabase==2.3.4
requests==2.32.3